feat(install/inject) Add inject and deps parameter
Add inject parameter to install() and deps parameter to inject(). This makes
it possible to specify depenedencies as an array of strings when code is to be
obfuscated.

BREAKING CHANGE: Property names are no longer automatically inferred as dependency
names. To work around this wrap your object in a function that accepts your
dependencies. Only functions can have dependencies injected.